app/requirements.txt
exchangelib==4.8.0
pywin32==304
PyYAML==6.0
psycopg2==2.9.4
SQLAlchemy==1.4.42
pandas==2.0.0
openpyxl==3.1.2
pyrfc==2.5.0
azure-ai-formrecognizer==3.3.0
azure-storage-blob==12.17.0